Moved courseStorage close to course module
